Tarefa1_2022li1g055

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

Função Principal:

A funçao mapaValido combina todas as funções auxiliares para verificar se o dado mapa é válido ou não

obstaculosValidos :: [(Terreno, [Obstaculo])] -> Bool

Função Auxiliar:

Estas funçoes verificam para cada terreno se os obstaculos que estao lá são válidos Cada uma das funçoes recebe um Terreno e uma lista de Obstaculos e verifca quais os obstaculos naquele terreno, sendo verdade caso respeite os parametros de cada uma.

relvaValida :: (Terreno, [Obstaculo]) -> Bool

estradaValida :: (Terreno, [Obstaculo]) -> Bool

rioValido :: (Terreno, [Obstaculo]) -> Bool

riosContiguos :: [(Terreno, [Obstaculo])] -> Bool

Função Auxiliar:

A funçao riosContiguos recebe uma listacom terrenos e os seus respetivos obstaculos, para o caso de serem dois rios na cabeca da lista, a funçao multiplica as velocidades dos dois se der menor que 0 vai voltar a testar com o segundo elemnto e o resto da lista. Se nao acontecer o caso anterior a funçao avança a head da lista e comeca a testar os proximos elementos

maxTronco :: [(Terreno, [Obstaculo])] -> Bool

Função Auxiliar:

A função maxTronco recebe uma lista de terrenos e seus respetivos obstaculos, aplicando a funçao auxiliar troncoValido a uma lista de obstaculos, esta por sua vez usa outra auxiliar chamada troncoAux que serve para contar quantos troncos tem numa lista de obstaculos, caso conte mais de 5 troncos isto fara com que a funçao de False tudo que seja 5 ou menos dará verdade

troncoValido :: [Obstaculo] -> Bool

troncoAux :: Int -> Int -> [Obstaculo] -> Int

maxCarro :: [(Terreno, [Obstaculo])] -> Bool

Função Auxiliar:

A função maxCarro recebe uma lista de terrenos e seus respetivos obstaculos, aplicando a funçao auxiliar carroValido a uma lista de obstaculos, esta por sua vez usa outra auxiliar chamada carroAux que serve para contar quantos carros tem numa lista de obstaculos, caso conte mais de 3 carros isto fara com que a funçao de False tudo que seja 3 ou menos dará verdade

carroValido :: [Obstaculo] -> Bool

carroAux :: Int -> Int -> [Obstaculo] -> Int

espacoLivre :: [(Terreno, [Obstaculo])] -> Bool

Função Auxiliar:

A função espacoLivre recebe uma lista de terrenos e a sua respetiva lista de obstaculos, utilizando a funçao auxiliar espacoLivreAux, esta verifica se na lista de obstaculos existe um obstaculo que seja nenhum, logo se a funçao auxiliar se verificar verdadeira entao vai ser aplicado outra vez o mesmo processo para o resto do mapa

espacoLivreAux :: [Obstaculo] -> Bool

comprimento :: Mapa -> Bool

Função Auxiliar:

A funçao comprimento recebe um Mapa com os seus terrenos e os respetivos obstaculos de cada terreno, a funçao pega na lista de obstaculos e verifica se o comprimento dessa lista é igual a largura do mapa caso seja igual vai continuar a testar para o resto do Mapa

maxTerrenos :: Mapa -> Bool

Função Auxiliar:

As tres funçoes numTerrenoContiguos recebem uma lista com terrenos e os seus respetivos obstaculos, estas tres funcoes servem como auxiliares a principal maxTerrenos, pois estas tres cada uma independente serve para contar quantos rios, estradas, relvas existem na lista que foi dada a funçao. Já só depois na principal applica se as tres auxiliares e caso as tres auxiliares respeitem as condiçoes impostas o Mapa será verdade de outra forma nao

numRiosContiguos :: [(Terreno, [Obstaculo])] -> Int -> Int

numEstradasContiguas :: [(Terreno, [Obstaculo])] -> Int -> Int

numRelvasContiguas :: [(Terreno, [Obstaculo])] -> Int -> Int