Tarefa2_2022li1g055

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

Funçao Principal:

A funçao estendeMapa receb um Mapa e um inteiro de 0 a 100, a funçao vai gerar uma nova linha do mapa e dependendo do numero escolhido poderá gerar um rio, estrada ou relva, isto tudo dentro dos parametros definidos nas funçoes auxiliares

proximosTerrenosValidos :: Mapa -> [Terreno]

Funçao Auxiliar:

A funçao proximosTerrenosValidos recebe um mapa com os seus respetivos terrenos e obstaculos, esta funçao vai verificar os ultimos terrenos da lista ou seja a sua head e caso esteja dentro de um dos casos definidos na função como por exemplo se os ultimos terrenos forem 4 Rios os proximos terrenos só podeem ser Estrada ou Relva no caso de serem 4 Estradas só pode ser Rio ou Relva, ou se os ultimos terrenos forem 4 Relvas só podem ser um Rio ou Estrada, no caso de os ultimos terrenos ou a head da lista nao se enquadrar em nenhum dos parametros anteriores os proximos terrenos válidos poderão ser qualquer um.

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

Funçao Auxiliar:

A funçao proximosObstaculosValidos recebe um inteiro, que representa a largura do mapa, e recebe tambem um terreno com a sua lista de obstaculos, esta função depnde do terreno que será, no caso de o terreno ser um Rio a função verifica o comprimento da lista de obstaculos se esta for igual a largura do mapa entao da lista vazia, ou seja, nao pode ser adicionado nenhum obstaculo, no caso do comprimento da lista de obstaculos ser menor que a largura a funçao utiliza a funçao auxiliar espacoLivreAux, que verifica se existe algum obstaculo Nenhum na lista caso isto seja falso o unico obstaculo possivel de ser adicionado será Nenhum, se ja existir um obstaculo Nenhum a funçao usará outra funçao auxiliar troncoValido que verifica se a quantidade de troncos na funçao é válida quando se adiciona um Tronco a lista de obstaculos que a funçao recebe caso isto seja verdade os obstaculos válidos serao Nenhum e Tronco. Sendo este mesmo processo para os dois outros terrenos.

escolheTerreno :: [Terreno] -> Int -> Terreno

Funçao Auxiliar:

A funçao escolheTerreno recebe uma lista de terrenos e um inteiro, pega no comprimento da lista de terrenos tira um e multiplica pelo inteiro, divide isto tudo por 100 e depois arredonda com o numero que dá procura essa posição na lista inicial de terrenos, dando o terreno que se encontra nessa posiçao

adicionaVelocidade :: Terreno -> Terreno -> Int -> Terreno

Funçao Auxiliar:

A funçao adicionaVelocidade recebe dois terrenos e um inteiro, caso os terrenos sejam dois rios e a velocidade do segundo for maior que 0 entao a funçao vai adicionar uma velocidade negativa ao primeiro rio isto de acordo com o inteiro que a funçao receber. Dependendo do n(Int) a velociade dos terrenos vai ser diferente

escolheObstaculo :: [Obstaculo] -> Int -> Obstaculo

Funçao Auxiliar:

A funçao escolheObstaculo recebe uma lista de obstaculos e um inteiro, pega no inteiro dado e no comprimento da lista de obstaculos faz a sua divisao e devolve o seu resto sendo sempre 1 ou 0, depois procura na lista de obstaculo que corresponde a [Nenhum,ArvoreTroncoCarro] se o resto for 0 devolve Nenhum, se for 1 devolve um daqueles tres obstaculos dependendo de qual o terreno.

constroiObstaculos :: Int -> (Terreno, [Obstaculo]) -> Int -> [Obstaculo]

Funçao Auxiliar:

A funçao constroiObstaculos recebe um inteiro que corresponde a largura, recebe um terreno e a sua lista de obstaculos e ainda recebe outro inteiro, primeiro a funçao testa se a largura do Mapa é igual ao comprimento da lista de obstaculos, caso isto se verifique vai nos dar a mesma lista de obstaculos não adicionando outro. Para o caso de isto nao se verificar vai ser usada a funçao proximosObstaculosValidos para verificar que obstaculos podem ser adicionados depois utiliza a funçao escolheObstaculo para escolher um obstaculo, este obstaculo é adicionado á lista anterior de obstaculos repetindo este processo até que a lista de obstaculos tenha o mesmo comprimento da largura do Mapa.