Tarefa4_2022li1g055

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou :: Jogo -> Bool

Função Principal:

A funçao jogoTerminou aplica a função auxiliar jogoTerminouAux ás coordenadas em que o jogador se encontra e á largura do mapa; utiliza tambem a funçao auxiliar terrenoNasCoordenadas aplicada á lista de terrenos e obstaculos do mapa dado e as coordenadas do jogador.

jogoTerminouAux :: Coordenadas -> Int -> (Terreno, [Obstaculo]) -> Bool

Funçao Auxiliar:

A Funçao jogoTerminouAux recebe umas coordenadas do jogador, um inteiro que corresponderá a largura do mapa e ainda um terreno e os seus respetivos obstaculos; quando o terreno for relva e a coordenada em x for menor que 0, maior ou igual a largura do mapa corresponde a True, ou seja que o jogo terminou; no caso do terreno ser Rio e a coordenada em x for menor que 0, maior ou igual a largura do mapa ou ainda a coordenada x corresponder a um obstaculo Nenhum, isto significa que o jogo terminou; por fim se o terreno for Estrada e mais uma vez a coordenada em x for menor que 0, maior ou igual a largura do mapa ou ainda a coordenada x corresponder a um obstaculo Carro, isto significa que o jogo terminou.

terrenoNasCoordenadas :: [(Terreno, [Obstaculo])] -> Coordenadas -> (Terreno, [Obstaculo])

Funçao Auxiliar:

A funçao terrenoNasCoordenadas recebe uma lista de terrenos e seus respetivos obstaculos e umas coordenadas, esta funçao vai procurar que terreno se encontra nas coordenadas dadas e é isso que a funçao nos vai dar, isto é, o terreno que se encontra naquelas coordenadas.

elemIndices :: Eq a => a -> [a] -> [Int]

Funçao Auxiliar:

A funçao elemIndices calcula a lista de posições em que um dado elemento ocorre numa lista.